requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
transformers>=4.43.3
pillow>=10.2.0
matplotlib>=3.8.0
gradio>=4.20.0
